Finished up main tasks

+++ src/com/itacademy/eshop/product/Product.java
@@ -48,7 +48,8 @@ public class Product {
         return rating / reviews.size();
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
+
 }
